Allow node server to define initial configuration.
Normally, when a node server is started by Polyglot, Polyglot
sends the configuration (mqtt host/port and profile num) via
stdin. The init() function will read from stdin to get the
configuration. When testing, it may be desirable to strt the
node server directly from the command line. The init() function
also takes an optional structure, with the host/port/profile
information and will use that instead of trying to read it from
stdin.